yamdb/requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==3.2.23
django-debug-toolbar==4.3.0
django-filter==2.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
pluggy==1.4.0
py==1.11.0
PyJWT==2.8.0
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.11.0
uritemplate==4.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
